backend/requirements.txt
fastapi
httpx
uvicorn
scikit-learn
# For OpenRouter/Gemini LLM API
python-dotenv
pytest
numpy